refactor: densify leftover 380-line overlay, diplomacy, and 320dp hosts
Extract shared fixtures for Slice D of #4606 so the remaining listed
app/test hosts sit at or below 360 physical lines without dropping cases.
