+++ app/test/diplomacy_detail_screen_test.dart
@@ -304,80 +304,4 @@ void main() {
       expect(popEvents, hasLength(1));
     },
   );
-
-  group('current relation card chrome', () {
-    for (final c
-        in <
-          ({
-            String name,
-            Game Function() game,
-            void Function(WidgetTester tester) assertUi,
-          })
-        >[
-          (
-            name: 'War label in --danger colour',
-            game: () => diplomacyDetailMinimalGame(
-              eventType: DiplomaticEventType.declareWar,
-              atWar: true,
-            ),
-            assertUi: (tester) {
-              expect(find.text('CURRENT RELATION'), findsOneWidget);
-              final Text war = tester.widget(find.text('War'));
-              expect(war.style?.color, EditorialMonoclePalette.danger);
-            },
-          ),
-          (
-            name: 'Peace label in --success colour',
-            game: diplomacyDetailMinimalGame,
-            assertUi: (tester) {
-              expect(find.text('CURRENT RELATION'), findsOneWidget);
-              final Text peace = tester.widget(find.text('Peace'));
-              expect(peace.style?.color, EditorialMonoclePalette.success);
-            },
-          ),
-          (
-            name:
-                'ALLIANCE badge in --accent for formal alliance (Refs #3625 AC4)',
-            game: () => diplomacyDetailMinimalGame(
-              eventType: DiplomaticEventType.allianceFormed,
-              score: 90,
-              formalAlliance: true,
-            ),
-            assertUi: (tester) {
-              expect(find.text('CURRENT RELATION'), findsOneWidget);
-              final Finder badge = find.text(kDiplomacyAllianceBadgeLabel);
-              expect(badge, findsOneWidget);
-              final Text badgeText = tester.widget<Text>(badge);
-              expect(badgeText.style?.color, EditorialMonoclePalette.accent);
-              expect(kDiplomacyAllianceBadgeLabel, isNot('Friendly'));
-            },
-          ),
-          (
-            name:
-                'omits ALLIANCE badge for informal Allied band (Refs #3625 AC4 negative)',
-            game: () => diplomacyDetailMinimalGame(score: 90),
-            assertUi: (tester) {
-              expect(find.text('CURRENT RELATION'), findsOneWidget);
-              expect(find.text(kDiplomacyAllianceBadgeLabel), findsNothing);
-              expect(find.textContaining('Devoted'), findsOneWidget);
-            },
-          ),
-        ]) {
-      testWidgets(c.name, (WidgetTester tester) async {
-        final game = c.game();
-        final relation = getRelation(
-          game,
-          diplomacyDetailHumanId,
-          diplomacyDetailOtherId,
-        );
-        expect(relation, isNotNull);
-        await pumpDiplomacyDetailOtherGp(
-          tester,
-          game: game,
-          relation: relation,
-        );
-        c.assertUi(tester);
-      });
-    }
-  });
 }
