+++ app/test/main_menu_wood_panel_hover_test.dart
@@ -3,163 +3,11 @@
 // SPEC/ui/main-menu.md § Variant rendering hover ACs:
 import 'package:colonizethis_app_ui_chrome/config/editorial_monocle_palette.dart';
 import 'package:colonizethis_app/widgets/ct_nine_patch_button.dart';
-import 'package:colonizethis_app/widgets/main_menu.dart';
 import 'package:colonizethis_test/test.dart' show suppressLogsForTests;
-import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import 'app_shell_harness.dart';
-
-/// Pumps a `CtMainMenu` configured for the `pixelArt` variant under the shared
-/// editorial-monocle app shell so the wood-panel button hover state can be
-/// inspected through the inner `CtNinePatchButton` widget.
-Future<void> _pumpPixelArtMainMenu(WidgetTester tester) async {
-  await pumpAppShell(
-    tester,
-    settle: true,
-    child: CtMainMenu(
-      variant: MainMenuVariant.pixelArt,
-      state: MainMenuState.default_,
-      version: 'v1.0.0',
-      onQuickStart: () {},
-      onNewGame: () {},
-      onLoadGame: () {},
-      onSettings: () {},
-      onQuit: () {},
-    ),
-  );
-}
-
-/// Returns the `CtNinePatchButton` ancestor of the menu label [label].
-Finder _woodPanelButtonFor(String label) {
-  return find.ancestor(
-    of: find.text(label),
-    matching: find.byType(CtNinePatchButton),
-  );
-}
-
-/// Returns the gradient-painting `DecoratedBox` painted by the
-/// `CtNinePatchButton` whose label is [label]. The button paints exactly
-/// one such box (the surface), so the finder is unambiguous.
-DecoratedBox _findSurfaceBoxFor(WidgetTester tester, String label) {
-  final Finder boxes = find.descendant(
-    of: _woodPanelButtonFor(label),
-    matching: find.byWidgetPredicate(
-      (Widget w) =>
-          w is DecoratedBox &&
-          w.decoration is BoxDecoration &&
-          (w.decoration as BoxDecoration).gradient != null,
-    ),
-  );
-  expect(
-    boxes,
-    findsAtLeastNWidgets(1),
-    reason:
-        'Wood-panel button "$label" must paint a gradient surface DecoratedBox',
-  );
-  return tester.widget<DecoratedBox>(boxes.first);
-}
-
-/// Returns the engraved label `RichText` rendered inside the
-/// `CtNinePatchButton` whose label is [label].
-RichText _findLabelRichTextFor(WidgetTester tester, String label) {
-  return tester.widget<RichText>(
-    find.descendant(of: find.text(label), matching: find.byType(RichText)),
-  );
-}
-
-/// Returns the `CtNinePatchButtonBracketsPainter` instance currently driving
-/// the brass-bracket overlay of the `CtNinePatchButton` whose label is [label].
-CustomPainter _findBrassPainterFor(WidgetTester tester, String label) {
-  final Finder painters = find.descendant(
-    of: _woodPanelButtonFor(label),
-    matching: find.byWidgetPredicate(
-      (Widget w) =>
-          w is CustomPaint &&
-          w.painter != null &&
-          w.painter.runtimeType.toString() == 'CtNinePatchButtonBracketsPainter',
-    ),
-  );
-  expect(
-    painters,
-    findsOneWidget,
-    reason:
-        'Wood-panel button "$label" must paint a single brass-corner-brackets '
-        'overlay (CustomPaint with CtNinePatchButtonBracketsPainter).',
-  );
-  final CustomPaint paint = tester.widget<CustomPaint>(painters);
-  return paint.painter!;
-}
-
-/// Reads the `color` field off the brass-bracket painter via `toString()`.
-/// The painter is private (`CtNinePatchButtonBracketsPainter` in
-/// `ct_nine_patch_button_brackets.dart`) and exposes its color only through its
-/// `paint(...)` call; we parse the painter's debug string which embeds the
-/// `color: Color(...)` value the painter holds. Falls back to the painted
-/// canvas only if the debug string does not encode the color.
-///
-/// Returning a `Color?` lets the test assert the color shifted, even though
-/// we cannot reach the private field directly.
-Color? _brassColorFromPainter(CustomPainter painter) {
-  // The painter's runtimeType already pins the class identity above; we now
-  // inspect the painter's `toString()` debug representation, which Dart
-  // generates from the class name. Since `CtNinePatchButtonBracketsPainter`
-  // overrides `shouldRepaint` but not `toString`, the default representation
-  // is just the class name — which means we cannot extract the color this
-  // way. Instead, paint into a recording canvas and capture the first paint
-  // call's color via a tiny shim.
-  final _ColorRecordingCanvas canvas = _ColorRecordingCanvas();
-  painter.paint(canvas, const Size(120, 48));
-  return canvas.firstColor;
-}
-
-/// Minimal `Canvas` shim that records the color of the first `drawRect`
-/// invocation. Sufficient for `CtNinePatchButtonBracketsPainter.paint`, which
-/// uses a single `Paint` instance for every rect.
-class _ColorRecordingCanvas implements Canvas {
-  Color? firstColor;
-
-  @override
-  void drawRect(Rect rect, Paint paint) {
-    firstColor ??= paint.color;
-  }
-
-  @override
-  dynamic noSuchMethod(Invocation invocation) => null;
-}
-
-Future<TestGesture> _addMousePointer(WidgetTester tester) async {
-  final TestGesture gesture = await tester.createGesture(
-    kind: PointerDeviceKind.mouse,
-  );
-  addTearDown(gesture.removePointer);
-  await gesture.addPointer(location: const Offset(-1000, -1000));
-  await tester.pump();
-  return gesture;
-}
-
-/// Pumps enough frames for [CtNinePatchButton.animationDuration] (120 ms)
-/// hover color animation to finish. We intentionally avoid
-/// `pumpAndSettle()` here because the `_PixelArtButton` wood-panel wrapper
-/// drives an **infinite** bob animation (`_bobController.repeat(reverse:
-/// true)`) on hover-enter, and `pumpAndSettle()` would block indefinitely
-/// waiting for that infinite controller to stop.
-Future<void> _settleHoverColorAnimation(WidgetTester tester) async {
-  await tester.pump();
-  await tester.pump(CtNinePatchButton.animationDuration);
-  await tester.pump(const Duration(milliseconds: 16));
-}
-
-/// Pumps frames after a hover-exit. The wood-panel bob controller is
-/// `stop()` + `reset()`-ed inside the exit handler, so once the color
-/// animation has run for [CtNinePatchButton.animationDuration]
-/// `pumpAndSettle()` is safe again.
-Future<void> _settleHoverExitAnimation(WidgetTester tester) async {
-  await tester.pump();
-  await tester.pump(CtNinePatchButton.animationDuration);
-  await tester.pumpAndSettle();
-}
+import 'main_menu_wood_panel_hover_test_support.dart';
 
 void main() {
   suppressLogsForTests();
@@ -174,11 +22,9 @@ void main() {
         '--accent and brightens the brass corner brackets to --accent-bright '
         'at hover alpha (border accent strengthens + corners brighten)',
         (WidgetTester tester) async {
-          await _pumpPixelArtMainMenu(tester);
+          await pumpPixelArtMainMenu(tester);
 
-          // Rest-state baseline: border resolves to --border, brass painter
-          // paints --accent at the default 0.75 alpha.
-          final DecoratedBox restSurface = _findSurfaceBoxFor(
+          final DecoratedBox restSurface = findWoodPanelSurfaceBox(
             tester,
             'New Game',
           );
@@ -186,8 +32,8 @@ void main() {
               (restSurface.decoration as BoxDecoration).border! as Border;
           expect(restBorder.top.color, EditorialMonoclePalette.border);
 
-          final Color? restBrass = _brassColorFromPainter(
-            _findBrassPainterFor(tester, 'New Game'),
+          final Color? restBrass = brassColorFromPainter(
+            findWoodPanelBrassPainter(tester, 'New Game'),
           );
           expect(restBrass, isNotNull);
           expect(
@@ -203,16 +49,14 @@ void main() {
                 'CtNinePatchButton.defaultCornerAlpha (0.75)',
           );
 
-          // Drive a mouse pointer into the center of the New Game button so
-          // the inner CtNinePatchButton's MouseRegion fires onEnter.
-          final TestGesture gesture = await _addMousePointer(tester);
+          final TestGesture gesture = await addMainMenuMousePointer(tester);
           final Offset center = tester.getCenter(
-            _woodPanelButtonFor('New Game'),
+            woodPanelButtonFor('New Game'),
           );
           await gesture.moveTo(center);
-          await _settleHoverColorAnimation(tester);
+          await settleHoverColorAnimation(tester);
 
-          final DecoratedBox hoverSurface = _findSurfaceBoxFor(
+          final DecoratedBox hoverSurface = findWoodPanelSurfaceBox(
             tester,
             'New Game',
           );
@@ -226,8 +70,8 @@ void main() {
                 '--accent (mockup .menu-btn:hover { border-color: var(--accent) })',
           );
 
-          final Color? hoverBrass = _brassColorFromPainter(
-            _findBrassPainterFor(tester, 'New Game'),
+          final Color? hoverBrass = brassColorFromPainter(
+            findWoodPanelBrassPainter(tester, 'New Game'),
           );
           expect(hoverBrass, isNotNull);
           expect(
@@ -251,28 +95,32 @@ void main() {
         'label color to --accent-bright while preserving the 1 px downward '
         '--surface engrave shadow',
         (WidgetTester tester) async {
-          await _pumpPixelArtMainMenu(tester);
+          await pumpPixelArtMainMenu(tester);
 
-          // Rest-state baseline: engraved label resolves to --accent and
-          // carries exactly one 1 px downward --surface shadow.
-          final RichText restLabel = _findLabelRichTextFor(tester, 'New Game');
+          final RichText restLabel = findWoodPanelLabelRichText(
+            tester,
+            'New Game',
+          );
           final TextSpan restSpan = restLabel.text as TextSpan;
           expect(restSpan.style?.color, EditorialMonoclePalette.accent);
           final List<Shadow>? restShadows = restSpan.style?.shadows;
           expect(restShadows, isNotNull);
           expect(restShadows!.length, 1);
-          expect(restShadows.first.offset, CtNinePatchButton.engravedShadowOffset);
+          expect(
+            restShadows.first.offset,
+            CtNinePatchButton.engravedShadowOffset,
+          );
           expect(restShadows.first.blurRadius, 0);
           expect(restShadows.first.color, EditorialMonoclePalette.surface);
 
-          final TestGesture gesture = await _addMousePointer(tester);
+          final TestGesture gesture = await addMainMenuMousePointer(tester);
           final Offset center = tester.getCenter(
-            _woodPanelButtonFor('New Game'),
+            woodPanelButtonFor('New Game'),
           );
           await gesture.moveTo(center);
-          await _settleHoverColorAnimation(tester);
+          await settleHoverColorAnimation(tester);
 
-          final RichText hoverLabel = _findLabelRichTextFor(
+          final RichText hoverLabel = findWoodPanelLabelRichText(
             tester,
             'New Game',
           );
@@ -315,29 +163,26 @@ void main() {
         'wood-panel button reverts border to --border, brass corners to '
         '--accent × 0.75 alpha, and label color back to --accent',
         (WidgetTester tester) async {
-          await _pumpPixelArtMainMenu(tester);
+          await pumpPixelArtMainMenu(tester);
 
-          final TestGesture gesture = await _addMousePointer(tester);
+          final TestGesture gesture = await addMainMenuMousePointer(tester);
           final Offset center = tester.getCenter(
-            _woodPanelButtonFor('New Game'),
+            woodPanelButtonFor('New Game'),
           );
           await gesture.moveTo(center);
-          await _settleHoverColorAnimation(tester);
+          await settleHoverColorAnimation(tester);
 
-          // Sanity: hover state was reached before we move off.
           final Border hoverBorder =
-              (_findSurfaceBoxFor(tester, 'New Game').decoration
-                      as BoxDecoration)
-                  .border!
-              as Border;
+              (findWoodPanelSurfaceBox(tester, 'New Game').decoration
+                          as BoxDecoration)
+                      .border!
+                  as Border;
           expect(hoverBorder.top.color, EditorialMonoclePalette.accent);
 
-          // Move the pointer well outside the menu. The bob controller is
-          // stopped+reset on exit so pumpAndSettle() is safe again.
           await gesture.moveTo(const Offset(-1000, -1000));
-          await _settleHoverExitAnimation(tester);
+          await settleHoverExitAnimation(tester);
 
-          final DecoratedBox restSurface = _findSurfaceBoxFor(
+          final DecoratedBox restSurface = findWoodPanelSurfaceBox(
             tester,
             'New Game',
           );
@@ -351,8 +196,8 @@ void main() {
                 '--border (hover is strictly transient).',
           );
 
-          final Color? restBrass = _brassColorFromPainter(
-            _findBrassPainterFor(tester, 'New Game'),
+          final Color? restBrass = brassColorFromPainter(
+            findWoodPanelBrassPainter(tester, 'New Game'),
           );
           expect(restBrass, isNotNull);
           expect(
@@ -370,7 +215,10 @@ void main() {
                 'CtNinePatchButton.defaultCornerAlpha (0.75)',
           );
 
-          final RichText restLabel = _findLabelRichTextFor(tester, 'New Game');
+          final RichText restLabel = findWoodPanelLabelRichText(
+            tester,
+            'New Game',
+          );
           final TextSpan restSpan = restLabel.text as TextSpan;
           expect(
             restSpan.style?.color,
