+++ app/test/province_offer_peace_shortcut_test.dart
@@ -1,116 +1,32 @@
 // Pins MAP20001 Political owner standing + Offer Peace (Refs #4479).
 // SPEC: SPEC/ui/province-sea-zone-detail-overlay.md — Political standing / Offer Peace.
-import 'package:colonizethis_app/features/game/flame/map_state/game_map_area_province_action_states_offer_peace.dart';
 import 'package:colonizethis_app/features/game/flame/map_state/game_map_area_state_logic.dart';
 import 'package:colonizethis_app/features/game/widgets/diplomacy/diplomacy_panel_constants.dart'
     show kDiplomacyAllianceBadgeLabel;
 import 'package:colonizethis_app_l10n/l10n/app_localizations_en.dart';
-import 'package:colonizethis_data/colonizethis_data.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:colonizethis_test/test.dart' show suppressLogsForTests;
-import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import 'province_detail_overlay_host_support_fixtures.dart';
+import 'province_offer_peace_shortcut_test_support.dart';
 import 'province_overlay_test_harness.dart';
-
-const String _kGameId = 'g_offer_peace_shortcut';
-const String _kHumanPlayerId = 'gp1';
-const String _kRivalId = 'gp2';
-const String _kProvinceId = 'oldWorld|p_rival';
-const String _kTileKey = 'oldWorld|p_rival|0|0';
-
-final MapTopology _topology = MapTopology(
-  nodes: const [
-    TopologyNode(
-      id: 'oldWorld|p_rival',
-      regionId: 'oldWorld',
-      type: TopologyNodeType.province,
-    ),
-    TopologyNode(
-      id: 'oldWorld|p_human',
-      regionId: 'oldWorld',
-      type: TopologyNodeType.province,
-    ),
-  ],
-  edges: const [],
-);
-
-Game _buildGame({
-  required String ownerId,
-  RelationState relationState = RelationState.atPeace,
-  bool formalAlliance = false,
-  bool includeRelation = true,
-}) {
-  return Game(
-    id: _kGameId,
-    worldState: WorldState(
-      turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 1),
-      oldWorld: RegionData(
-        provinces: [
-          Province(
-            id: _kProvinceId,
-            regionId: 'oldWorld',
-            ownerId: ownerId,
-            townTileKey: _kTileKey,
-          ),
-          const Province(
-            id: 'oldWorld|p_human',
-            regionId: 'oldWorld',
-            ownerId: _kHumanPlayerId,
-            townTileKey: 'oldWorld|p_human|0|0',
-          ),
-        ],
-        units: const [],
-      ),
-      newWorld: const RegionData(provinces: [], units: []),
-    ),
-    players: const [
-      Player(
-        id: _kHumanPlayerId,
-        displayName: 'Human',
-        isHuman: true,
-        capitalProvinceId: 'oldWorld|p_human',
-        treasury: 5000,
-      ),
-      Player(
-        id: _kRivalId,
-        displayName: 'Rival',
-        isHuman: false,
-        capitalProvinceId: _kProvinceId,
-      ),
-    ],
-    minorNations: const [],
-    tribes: const [],
-    diplomacyRelations: includeRelation
-        ? [
-            DiplomacyRelation(
-              factionId1: _kHumanPlayerId,
-              factionId2: _kRivalId,
-              state: relationState,
-              formalAlliance: formalAlliance,
-              score: formalAlliance ? 90 : 50,
-            ),
-          ]
-        : const [],
-  );
-}
 
 void main() {
   suppressLogsForTests();
 
   group('Owner standing / Offer Peace action state', () {
     test('at war shows standing and Offer Peace when probe accepts', () {
-      final game = _buildGame(
-        ownerId: _kRivalId,
+      final game = offerPeaceBuildGame(
+        ownerId: offerPeaceRivalId,
         relationState: RelationState.atWar,
       );
       final state =
           GameMapAreaStateLogicProvinceActions.provinceOfferPeaceActionState(
             game: game,
-            humanPlayerId: _kHumanPlayerId,
-            provinceId: _kProvinceId,
-            topology: _topology,
+            humanPlayerId: offerPeaceHumanPlayerId,
+            provinceId: offerPeaceProvinceId,
+            topology: offerPeaceTopology,
             currentOrders: const Orders(),
             isSeaZone: false,
           );
@@ -123,12 +39,12 @@ void main() {
     });
 
     test('at peace shows standing without Offer Peace', () {
-      final game = _buildGame(ownerId: _kRivalId);
+      final game = offerPeaceBuildGame(ownerId: offerPeaceRivalId);
       final state = GameMapAreaProvinceActionStatesOfferPeace.compute(
         game: game,
-        humanPlayerId: _kHumanPlayerId,
-        provinceId: _kProvinceId,
-        topology: _topology,
+        humanPlayerId: offerPeaceHumanPlayerId,
+        provinceId: offerPeaceProvinceId,
+        topology: offerPeaceTopology,
         currentOrders: const Orders(),
         isSeaZone: false,
       );
@@ -139,12 +55,15 @@ void main() {
     });
 
     test('formal alliance shows ALLIANCE badge at peace', () {
-      final game = _buildGame(ownerId: _kRivalId, formalAlliance: true);
+      final game = offerPeaceBuildGame(
+        ownerId: offerPeaceRivalId,
+        formalAlliance: true,
+      );
       final state = GameMapAreaProvinceActionStatesOfferPeace.compute(
         game: game,
-        humanPlayerId: _kHumanPlayerId,
-        provinceId: _kProvinceId,
-        topology: _topology,
+        humanPlayerId: offerPeaceHumanPlayerId,
+        provinceId: offerPeaceProvinceId,
+        topology: offerPeaceTopology,
         currentOrders: const Orders(),
         isSeaZone: false,
       );
@@ -154,12 +73,12 @@ void main() {
     });
 
     test('hides when own province', () {
-      final game = _buildGame(ownerId: _kHumanPlayerId);
+      final game = offerPeaceBuildGame(ownerId: offerPeaceHumanPlayerId);
       final state = GameMapAreaProvinceActionStatesOfferPeace.compute(
         game: game,
-        humanPlayerId: _kHumanPlayerId,
+        humanPlayerId: offerPeaceHumanPlayerId,
         provinceId: 'oldWorld|p_human',
-        topology: _topology,
+        topology: offerPeaceTopology,
         currentOrders: const Orders(),
         isSeaZone: false,
       );
@@ -168,12 +87,15 @@ void main() {
     });
 
     test('hides when no DiplomacyRelation', () {
-      final game = _buildGame(ownerId: _kRivalId, includeRelation: false);
+      final game = offerPeaceBuildGame(
+        ownerId: offerPeaceRivalId,
+        includeRelation: false,
+      );
       final state = GameMapAreaProvinceActionStatesOfferPeace.compute(
         game: game,
-        humanPlayerId: _kHumanPlayerId,
-        provinceId: _kProvinceId,
-        topology: _topology,
+        humanPlayerId: offerPeaceHumanPlayerId,
+        provinceId: offerPeaceProvinceId,
+        topology: offerPeaceTopology,
         currentOrders: const Orders(),
         isSeaZone: false,
       );
@@ -182,15 +104,15 @@ void main() {
     });
 
     test('hides when sea-zone context', () {
-      final game = _buildGame(
-        ownerId: _kRivalId,
+      final game = offerPeaceBuildGame(
+        ownerId: offerPeaceRivalId,
         relationState: RelationState.atWar,
       );
       final state = GameMapAreaProvinceActionStatesOfferPeace.compute(
         game: game,
-        humanPlayerId: _kHumanPlayerId,
-        provinceId: _kProvinceId,
-        topology: _topology,
+        humanPlayerId: offerPeaceHumanPlayerId,
+        provinceId: offerPeaceProvinceId,
+        topology: offerPeaceTopology,
         currentOrders: const Orders(),
         isSeaZone: true,
       );
@@ -199,21 +121,21 @@ void main() {
     });
 
     test('pending offerPeace enables Cancel path', () {
-      final game = _buildGame(
-        ownerId: _kRivalId,
+      final game = offerPeaceBuildGame(
+        ownerId: offerPeaceRivalId,
         relationState: RelationState.atWar,
       );
       final state = GameMapAreaProvinceActionStatesOfferPeace.compute(
         game: game,
-        humanPlayerId: _kHumanPlayerId,
-        provinceId: _kProvinceId,
-        topology: _topology,
+        humanPlayerId: offerPeaceHumanPlayerId,
+        provinceId: offerPeaceProvinceId,
+        topology: offerPeaceTopology,
         currentOrders: const Orders(
           diplomaticOrdersByPlayerId: {
-            _kHumanPlayerId: [
+            offerPeaceHumanPlayerId: [
               DiplomaticOrder(
                 type: DiplomaticOrderType.offerPeace,
-                targetFactionId: _kRivalId,
+                targetFactionId: offerPeaceRivalId,
               ),
             ],
           },
@@ -231,23 +153,22 @@ void main() {
       final bus = AppEventBus();
       const order = DiplomaticOrder(
         type: DiplomaticOrderType.offerPeace,
-        targetFactionId: _kRivalId,
+        targetFactionId: offerPeaceRivalId,
       );
-      final game = _buildGame(
-        ownerId: _kRivalId,
+      final game = offerPeaceBuildGame(
+        ownerId: offerPeaceRivalId,
         relationState: RelationState.atWar,
       );
-      final confirmFuture = bus
-          .on<ConfirmDialogEvent>()
-          .first
-          .timeout(const Duration(seconds: 2));
+      final confirmFuture = bus.on<ConfirmDialogEvent>().first.timeout(
+        const Duration(seconds: 2),
+      );
       final appendFuture = bus
           .on<AppendDiplomaticOrderRequestedEvent>()
           .first
           .timeout(const Duration(seconds: 2));
       final callbacks = provinceDetailCallbacks(
         game: game,
-        selectedTileKey: _kTileKey,
+        selectedTileKey: offerPeaceTileKey,
         exploreEnabled: false,
         prospectEnabled: false,
         buildImprovementEnabled: false,
@@ -255,12 +176,12 @@ void main() {
         buildFortEnabled: false,
         buildPortEnabled: false,
         purchaseLandEnabled: false,
-        provinceId: _kProvinceId,
+        provinceId: offerPeaceProvinceId,
         offerPeaceEnabled: true,
         offerPeacePending: false,
         offerPeaceOrder: order,
         offerPeaceTargetName: 'Rival',
-        combinedTopology: _topology,
+        combinedTopology: offerPeaceTopology,
         bus: bus,
       );
       expect(callbacks.onOfferPeaceTap, isNotNull);
@@ -275,10 +196,10 @@ void main() {
       final bus = AppEventBus();
       const order = DiplomaticOrder(
         type: DiplomaticOrderType.offerPeace,
-        targetFactionId: _kRivalId,
+        targetFactionId: offerPeaceRivalId,
       );
-      final game = _buildGame(
-        ownerId: _kRivalId,
+      final game = offerPeaceBuildGame(
+        ownerId: offerPeaceRivalId,
         relationState: RelationState.atWar,
       );
       final removeFuture = bus
@@ -287,7 +208,7 @@ void main() {
           .timeout(const Duration(seconds: 2));
       final callbacks = provinceDetailCallbacks(
         game: game,
-        selectedTileKey: _kTileKey,
+        selectedTileKey: offerPeaceTileKey,
         exploreEnabled: false,
         prospectEnabled: false,
         buildImprovementEnabled: false,
@@ -304,22 +225,22 @@ void main() {
       callbacks.onOfferPeaceTap!();
       final remove = await removeFuture;
       expect(remove.type, DiplomaticOrderType.offerPeace);
-      expect(remove.targetFactionId, _kRivalId);
+      expect(remove.targetFactionId, offerPeaceRivalId);
     });
   });
 
   group('Political standing / Offer Peace widget', () {
     testWidgets('renders At war and Offer Peace when shown', (tester) async {
-      final game = _buildGame(
-        ownerId: _kRivalId,
+      final game = offerPeaceBuildGame(
+        ownerId: offerPeaceRivalId,
         relationState: RelationState.atWar,
       );
       final l10n = AppLocalizationsEn();
       await tester.pumpWidget(
         buildProvinceOverlayDarkThemeShell(
           game: game,
-          displayId: _kProvinceId,
-          selectedTileKey: _kTileKey,
+          displayId: offerPeaceProvinceId,
+          selectedTileKey: offerPeaceTileKey,
           omniscientDetail: true,
           showOwnerStanding: true,
           ownerStandingAtWar: true,
@@ -329,7 +250,10 @@ void main() {
         ),
       );
       await tester.pump();
-      expect(find.text(l10n.provinceOverlay_ownerStandingAtWar), findsOneWidget);
+      expect(
+        find.text(l10n.provinceOverlay_ownerStandingAtWar),
+        findsOneWidget,
+      );
       expect(find.text(l10n.provinceOverlay_offerPeaceAction), findsOneWidget);
       expect(find.text('AT_WAR'), findsNothing);
     });
@@ -337,13 +261,16 @@ void main() {
     testWidgets('renders At peace without Offer Peace; ALLIANCE when allied', (
       tester,
     ) async {
-      final game = _buildGame(ownerId: _kRivalId, formalAlliance: true);
+      final game = offerPeaceBuildGame(
+        ownerId: offerPeaceRivalId,
+        formalAlliance: true,
+      );
       final l10n = AppLocalizationsEn();
       await tester.pumpWidget(
         buildProvinceOverlayDarkThemeShell(
           game: game,
-          displayId: _kProvinceId,
-          selectedTileKey: _kTileKey,
+          displayId: offerPeaceProvinceId,
+          selectedTileKey: offerPeaceTileKey,
           omniscientDetail: true,
           showOwnerStanding: true,
           ownerStandingAtWar: false,
@@ -361,16 +288,16 @@ void main() {
     });
 
     testWidgets('pending shows Cancel', (tester) async {
-      final game = _buildGame(
-        ownerId: _kRivalId,
+      final game = offerPeaceBuildGame(
+        ownerId: offerPeaceRivalId,
         relationState: RelationState.atWar,
       );
       final l10n = AppLocalizationsEn();
       await tester.pumpWidget(
         buildProvinceOverlayDarkThemeShell(
           game: game,
-          displayId: _kProvinceId,
-          selectedTileKey: _kTileKey,
+          displayId: offerPeaceProvinceId,
+          selectedTileKey: offerPeaceTileKey,
           omniscientDetail: true,
           showOwnerStanding: true,
           ownerStandingAtWar: true,
