+++ app/test/province_sea_zone_resource_labels_test.dart
@@ -1,86 +1,14 @@
 // Widget tests for resource icon + label in province overlay (Tile + Economic).
 // SPEC/ui/province-sea-zone-detail-overlay.md, SPEC/ui/pixel-art-ui-catalog.md.
 
-import 'package:colonizethis_logic/colonizethis_logic.dart'
-    show PlayerView, VisibilityLevel;
 import 'package:colonizethis_map/colonizethis_map.dart';
-import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:colonizethis_test/test.dart' show suppressLogsForTests;
 import 'package:flutter_test/flutter_test.dart';
 
 import 'package:colonizethis_app/widgets/resource_icon.dart';
 import 'package:colonizethis_app/widgets/strict_asset_icon.dart';
 
-import 'province_overlay_test_harness.dart';
-
-const _regionId = 'oldWorld';
-const _localProvinceId = 'pResTest';
-String get _fullProvinceId => '$_regionId|$_localProvinceId';
-
-String _tileKey(int x, int y) => '$_fullProvinceId|$x|$y';
-
-RegionMapViewData _regionWithCells(List<CellViewData> cells, int w, int h) {
-  return RegionMapViewData(
-    regionId: _regionId,
-    width: w,
-    height: h,
-    cellSize: 32,
-    cells: cells,
-    capitalMarkers: const [],
-    portMarkers: const [],
-    factionColors: const {},
-    greatPowerFactionIds: const {'gp1'},
-    terrainColors: const {},
-  );
-}
-
-Game _minimalGame({
-  required Map<String, List<String>> tileKeysByProvince,
-  Map<String, String> resourceByTileKey = const {},
-  Map<String, Map<String, String>> playerVisibilityByTile = const {},
-  Map<String, Set<String>> playerProspectedTiles = const {},
-}) {
-  return Game(
-    id: 'res_label_test',
-    worldState: WorldState(
-      turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 0),
-      oldWorld: RegionData(
-        provinces: [
-          Province(
-            id: _fullProvinceId,
-            regionId: _regionId,
-            displayName: 'ResTest',
-          ),
-        ],
-      ),
-      newWorld: const RegionData(),
-      tileKeysByRegionAndProvince: {_regionId: tileKeysByProvince},
-      resourceByTileKey: resourceByTileKey,
-      playerVisibilityByTile: playerVisibilityByTile,
-      playerProspectedTiles: playerProspectedTiles,
-    ),
-    players: const [
-      Player(id: 'gp1', displayName: 'Human', isHuman: true, treasury: 0),
-    ],
-  );
-}
-
-PlayerView _omniscientViewForTiles(Iterable<String> keys) {
-  return PlayerView(
-    playerId: 'gp1',
-    player: const Player(
-      id: 'gp1',
-      displayName: 'Human',
-      isHuman: true,
-      treasury: 0,
-    ),
-    ownUnitsById: const {},
-    provincesById: const {},
-    visibilityByTile: {for (final k in keys) k: VisibilityLevel.fullyVisible},
-    prospectedTiles: const {},
-    diplomacyByOtherId: const {},
-  );
-}
+import 'province_sea_zone_resource_labels_test_support.dart';
 
 void main() {
   suppressLogsForTests();
@@ -89,47 +17,39 @@ void main() {
     testWidgets(
       'Tile and Economic show ResourceLabelInline when grain is visible and prospected',
       (WidgetTester tester) async {
-        final tk = _tileKey(0, 0);
-        final game = _minimalGame(
-          tileKeysByProvince: {
-            _fullProvinceId: [tk],
-          },
-          resourceByTileKey: {tk: 'grain'},
-          playerVisibilityByTile: {
-            'gp1': {tk: 'fullyVisible'},
-          },
-          playerProspectedTiles: {
-            'gp1': {tk},
-          },
-        );
-        final region = _regionWithCells(
-          [
-            const CellViewData(
-              x: 0,
-              y: 0,
-              regionCellId: _localProvinceId,
-              isSea: false,
-              terrainTypeId: 'plains',
-              resourceId: 'grain',
-              visibility: TileVisibility.visible,
-            ),
-          ],
-          1,
-          1,
-        );
-
-        await tester.pumpWidget(
-          buildProvinceOverlayDarkThemeShell(
-            game: game,
-            region: region,
-            displayId: _fullProvinceId,
-            selectedTileKey: tk,
-            humanPlayerId: 'gp1',
-            playerView: _omniscientViewForTiles([tk]),
-            shellWidth: 800,
+        final tk = resourceLabelTileKey(0, 0);
+        await pumpResourceLabelOverlay(
+          tester,
+          game: resourceLabelMinimalGame(
+            tileKeysByProvince: {
+              resourceLabelFullProvinceId: [tk],
+            },
+            resourceByTileKey: {tk: 'grain'},
+            playerVisibilityByTile: {
+              'gp1': {tk: 'fullyVisible'},
+            },
+            playerProspectedTiles: {
+              'gp1': {tk},
+            },
           ),
+          region: resourceLabelRegionWithCells(
+            [
+              const CellViewData(
+                x: 0,
+                y: 0,
+                regionCellId: resourceLabelLocalProvinceId,
+                isSea: false,
+                terrainTypeId: 'plains',
+                resourceId: 'grain',
+                visibility: TileVisibility.visible,
+              ),
+            ],
+            1,
+            1,
+          ),
+          selectedTileKey: tk,
+          viewTileKeys: [tk],
         );
-        await tester.pumpAndSettle();
 
         expect(find.text('Grain'), findsNWidgets(2));
         expect(find.byType(ResourceLabelInline), findsNWidgets(2));
@@ -140,45 +60,37 @@ void main() {
     testWidgets(
       'Economic excludes unprospected tile even when resource is visible in Tile section',
       (WidgetTester tester) async {
-        final tk = _tileKey(0, 0);
-        final game = _minimalGame(
-          tileKeysByProvince: {
-            _fullProvinceId: [tk],
-          },
-          resourceByTileKey: {tk: 'grain'},
-          playerVisibilityByTile: {
-            'gp1': {tk: 'fullyVisible'},
-          },
-          playerProspectedTiles: const {'gp1': <String>{}},
-        );
-        final region = _regionWithCells(
-          [
-            const CellViewData(
-              x: 0,
-              y: 0,
-              regionCellId: _localProvinceId,
-              isSea: false,
-              terrainTypeId: 'plains',
-              resourceId: 'grain',
-              visibility: TileVisibility.visible,
-            ),
-          ],
-          1,
-          1,
-        );
-
-        await tester.pumpWidget(
-          buildProvinceOverlayDarkThemeShell(
-            game: game,
-            region: region,
-            displayId: _fullProvinceId,
-            selectedTileKey: tk,
-            humanPlayerId: 'gp1',
-            playerView: _omniscientViewForTiles([tk]),
-            shellWidth: 800,
+        final tk = resourceLabelTileKey(0, 0);
+        await pumpResourceLabelOverlay(
+          tester,
+          game: resourceLabelMinimalGame(
+            tileKeysByProvince: {
+              resourceLabelFullProvinceId: [tk],
+            },
+            resourceByTileKey: {tk: 'grain'},
+            playerVisibilityByTile: {
+              'gp1': {tk: 'fullyVisible'},
+            },
+            playerProspectedTiles: const {'gp1': <String>{}},
           ),
+          region: resourceLabelRegionWithCells(
+            [
+              const CellViewData(
+                x: 0,
+                y: 0,
+                regionCellId: resourceLabelLocalProvinceId,
+                isSea: false,
+                terrainTypeId: 'plains',
+                resourceId: 'grain',
+                visibility: TileVisibility.visible,
+              ),
+            ],
+            1,
+            1,
+          ),
+          selectedTileKey: tk,
+          viewTileKeys: [tk],
         );
-        await tester.pumpAndSettle();
 
         expect(find.text('Grain'), findsOneWidget);
         expect(find.byType(ResourceLabelInline), findsOneWidget);
@@ -188,46 +100,38 @@ void main() {
     testWidgets(
       'Economic excludes prospected tiles with no discovered resource',
       (WidgetTester tester) async {
-        final tk = _tileKey(0, 0);
-        final game = _minimalGame(
-          tileKeysByProvince: {
-            _fullProvinceId: [tk],
-          },
-          resourceByTileKey: const {},
-          playerVisibilityByTile: {
-            'gp1': {tk: 'fullyVisible'},
-          },
-          playerProspectedTiles: {
-            'gp1': {tk},
-          },
-        );
-        final region = _regionWithCells(
-          [
-            const CellViewData(
-              x: 0,
-              y: 0,
-              regionCellId: _localProvinceId,
-              isSea: false,
-              terrainTypeId: 'plains',
-              visibility: TileVisibility.visible,
-            ),
-          ],
-          1,
-          1,
-        );
-
-        await tester.pumpWidget(
-          buildProvinceOverlayDarkThemeShell(
-            game: game,
-            region: region,
-            displayId: _fullProvinceId,
-            selectedTileKey: tk,
-            humanPlayerId: 'gp1',
-            playerView: _omniscientViewForTiles([tk]),
-            shellWidth: 800,
+        final tk = resourceLabelTileKey(0, 0);
+        await pumpResourceLabelOverlay(
+          tester,
+          game: resourceLabelMinimalGame(
+            tileKeysByProvince: {
+              resourceLabelFullProvinceId: [tk],
+            },
+            resourceByTileKey: const {},
+            playerVisibilityByTile: {
+              'gp1': {tk: 'fullyVisible'},
+            },
+            playerProspectedTiles: {
+              'gp1': {tk},
+            },
           ),
+          region: resourceLabelRegionWithCells(
+            [
+              const CellViewData(
+                x: 0,
+                y: 0,
+                regionCellId: resourceLabelLocalProvinceId,
+                isSea: false,
+                terrainTypeId: 'plains',
+                visibility: TileVisibility.visible,
+              ),
+            ],
+            1,
+            1,
+          ),
+          selectedTileKey: tk,
+          viewTileKeys: [tk],
         );
-        await tester.pumpAndSettle();
 
         expect(find.byType(ResourceLabelInline), findsNothing);
         expect(find.textContaining('Resource:'), findsWidgets);
@@ -237,43 +141,35 @@ void main() {
     testWidgets('Tile shows plain em dash when tile has no visible resource', (
       WidgetTester tester,
     ) async {
-      final tk = _tileKey(0, 0);
-      final game = _minimalGame(
-        tileKeysByProvince: {
-          _fullProvinceId: [tk],
-        },
-        resourceByTileKey: const {},
-        playerVisibilityByTile: {
-          'gp1': {tk: 'fullyVisible'},
-        },
-      );
-      final region = _regionWithCells(
-        [
-          const CellViewData(
-            x: 0,
-            y: 0,
-            regionCellId: _localProvinceId,
-            isSea: false,
-            terrainTypeId: 'plains',
-            visibility: TileVisibility.visible,
-          ),
-        ],
-        1,
-        1,
-      );
-
-      await tester.pumpWidget(
-        buildProvinceOverlayDarkThemeShell(
-          game: game,
-          region: region,
-          displayId: _fullProvinceId,
-          selectedTileKey: tk,
-          humanPlayerId: 'gp1',
-          playerView: _omniscientViewForTiles([tk]),
-          shellWidth: 800,
+      final tk = resourceLabelTileKey(0, 0);
+      await pumpResourceLabelOverlay(
+        tester,
+        game: resourceLabelMinimalGame(
+          tileKeysByProvince: {
+            resourceLabelFullProvinceId: [tk],
+          },
+          resourceByTileKey: const {},
+          playerVisibilityByTile: {
+            'gp1': {tk: 'fullyVisible'},
+          },
         ),
+        region: resourceLabelRegionWithCells(
+          [
+            const CellViewData(
+              x: 0,
+              y: 0,
+              regionCellId: resourceLabelLocalProvinceId,
+              isSea: false,
+              terrainTypeId: 'plains',
+              visibility: TileVisibility.visible,
+            ),
+          ],
+          1,
+          1,
+        ),
+        selectedTileKey: tk,
+        viewTileKeys: [tk],
       );
-      await tester.pumpAndSettle();
 
       expect(find.byType(ResourceLabelInline), findsNothing);
       expect(find.textContaining('Resource:'), findsWidgets);
@@ -282,44 +178,36 @@ void main() {
     testWidgets(
       'Prospect-required resource hidden until prospected (no ResourceLabelInline)',
       (WidgetTester tester) async {
-        final tk = _tileKey(0, 0);
-        final game = _minimalGame(
-          tileKeysByProvince: {
-            _fullProvinceId: [tk],
-          },
-          resourceByTileKey: {tk: 'iron'},
-          playerVisibilityByTile: {
-            'gp1': {tk: 'fullyVisible'},
-          },
-        );
-        final region = _regionWithCells(
-          [
-            const CellViewData(
-              x: 0,
-              y: 0,
-              regionCellId: _localProvinceId,
-              isSea: false,
-              terrainTypeId: 'hills',
-              resourceId: 'iron',
-              visibility: TileVisibility.visible,
-            ),
-          ],
-          1,
-          1,
-        );
-
-        await tester.pumpWidget(
-          buildProvinceOverlayDarkThemeShell(
-            game: game,
-            region: region,
-            displayId: _fullProvinceId,
-            selectedTileKey: tk,
-            humanPlayerId: 'gp1',
-            playerView: _omniscientViewForTiles([tk]),
-            shellWidth: 800,
+        final tk = resourceLabelTileKey(0, 0);
+        await pumpResourceLabelOverlay(
+          tester,
+          game: resourceLabelMinimalGame(
+            tileKeysByProvince: {
+              resourceLabelFullProvinceId: [tk],
+            },
+            resourceByTileKey: {tk: 'iron'},
+            playerVisibilityByTile: {
+              'gp1': {tk: 'fullyVisible'},
+            },
           ),
+          region: resourceLabelRegionWithCells(
+            [
+              const CellViewData(
+                x: 0,
+                y: 0,
+                regionCellId: resourceLabelLocalProvinceId,
+                isSea: false,
+                terrainTypeId: 'hills',
+                resourceId: 'iron',
+                visibility: TileVisibility.visible,
+              ),
+            ],
+            1,
+            1,
+          ),
+          selectedTileKey: tk,
+          viewTileKeys: [tk],
         );
-        await tester.pumpAndSettle();
 
         expect(find.byType(ResourceLabelInline), findsNothing);
         expect(find.text('iron'), findsNothing);
@@ -329,57 +217,49 @@ void main() {
     testWidgets('Economic lists two province resources with icons (sorted)', (
       WidgetTester tester,
     ) async {
-      final tk0 = _tileKey(0, 0);
-      final tk1 = _tileKey(1, 0);
-      final game = _minimalGame(
-        tileKeysByProvince: {
-          _fullProvinceId: [tk0, tk1],
-        },
-        resourceByTileKey: {tk0: 'timber', tk1: 'grain'},
-        playerVisibilityByTile: {
-          'gp1': {tk0: 'fullyVisible', tk1: 'fullyVisible'},
-        },
-        playerProspectedTiles: {
-          'gp1': {tk0, tk1},
-        },
-      );
-      final region = _regionWithCells(
-        [
-          const CellViewData(
-            x: 0,
-            y: 0,
-            regionCellId: _localProvinceId,
-            isSea: false,
-            terrainTypeId: 'hardwoodForest',
-            resourceId: 'timber',
-            visibility: TileVisibility.visible,
-          ),
-          const CellViewData(
-            x: 1,
-            y: 0,
-            regionCellId: _localProvinceId,
-            isSea: false,
-            terrainTypeId: 'plains',
-            resourceId: 'grain',
-            visibility: TileVisibility.visible,
-          ),
-        ],
-        2,
-        1,
-      );
-
-      await tester.pumpWidget(
-        buildProvinceOverlayDarkThemeShell(
-          game: game,
-          region: region,
-          displayId: _fullProvinceId,
-          selectedTileKey: tk0,
-          humanPlayerId: 'gp1',
-          playerView: _omniscientViewForTiles([tk0, tk1]),
-          shellWidth: 800,
+      final tk0 = resourceLabelTileKey(0, 0);
+      final tk1 = resourceLabelTileKey(1, 0);
+      await pumpResourceLabelOverlay(
+        tester,
+        game: resourceLabelMinimalGame(
+          tileKeysByProvince: {
+            resourceLabelFullProvinceId: [tk0, tk1],
+          },
+          resourceByTileKey: {tk0: 'timber', tk1: 'grain'},
+          playerVisibilityByTile: {
+            'gp1': {tk0: 'fullyVisible', tk1: 'fullyVisible'},
+          },
+          playerProspectedTiles: {
+            'gp1': {tk0, tk1},
+          },
         ),
+        region: resourceLabelRegionWithCells(
+          [
+            const CellViewData(
+              x: 0,
+              y: 0,
+              regionCellId: resourceLabelLocalProvinceId,
+              isSea: false,
+              terrainTypeId: 'hardwoodForest',
+              resourceId: 'timber',
+              visibility: TileVisibility.visible,
+            ),
+            const CellViewData(
+              x: 1,
+              y: 0,
+              regionCellId: resourceLabelLocalProvinceId,
+              isSea: false,
+              terrainTypeId: 'plains',
+              resourceId: 'grain',
+              visibility: TileVisibility.visible,
+            ),
+          ],
+          2,
+          1,
+        ),
+        selectedTileKey: tk0,
+        viewTileKeys: [tk0, tk1],
       );
-      await tester.pumpAndSettle();
 
       expect(find.text('Grain'), findsOneWidget);
       expect(find.text('Timber'), findsNWidgets(2));
